20180712-useacabeca-c:  Adicionado texto mostrando a vida
