+++ PROJETOS/20180712-useacabeca-c/blasteroids/include/blasteroids.h
@@ -1,5 +1,6 @@
 #include <allegro5/allegro.h>
 #include <allegro5/allegro_primitives.h>
+#include <allegro5/allegro_ttf.h>
 #include <stdbool.h>
 
 #include <blasteroids/blasteroids_types.h>
