+++ PROJETOS/20180712-useacabeca-c/blasteroids/include/blasteroids/collision.h
@@ -1,0 +1,6 @@
+#ifndef BLASTEROIDS_COLLISION
+#define BLASTEROIDS_COLLISION
+
+int is_collision(GameContext *ctx);
+
+#endif
