+++ PROJETOS/20180712-useacabeca-c/blasteroids/include/blasteroids/text.h
@@ -1,0 +1,6 @@
+#ifndef BLASTEROIDS_TEXT
+#define BLASTEROIDS_TEXT
+
+void draw_life(GameContext *ctx);
+
+#endif
